python/ranker/requirements.txt
flask
git+https://github.com/twmht/python-rocksdb.git
psycopg2-binary
numpy